Updated settings to be separate components Added experiemental settings
